requirements.txt
streamlit
langgraph
langchain-core
langchain-groq
langchain-text-splitters
transformers
torch
torchvision
yfinance
ddgs
edgartools
python-dotenv
defeatbeta-api
pandas